feat: Add function to clone and checkout repositories
This commit adds a new function called `clone_and_checkout` to the `.devcontainer/clone_repos.sh` script. This function is responsible for cloning and checking out branches of repositories. It takes two parameters: the repository owner and the repository name.

The function clones the repository using the provided owner and name, and then checks out the `develop` branch if it exists. If the `develop` branch is not found, it checks out the `main` branch instead. If neither branch is found, it stays on the default branch.

This function is used to clone and checkout a list of repositories specified in the script. Currently, only the `mountainash-utils-os` repository is being cloned and checked out.

This commit also adds a message to indicate that all repositories have been cloned and the appropriate branches have been checked out.